found logical flaw in user score update. previously, the reviewer score got updated. uploading a paper does not require ownerId now. ownerId is set to uploader id. owner id is set to author with the paper instantly while uploading. user trying to create review a paper multiple times now will redirect to update. sorry for these many changes.

+++ ResearchPaperService/src/main/java/com/ynm/researchpaperservice/Repository/AuthorRepository.java
@@ -14,5 +14,5 @@ public interface AuthorRepository extends JpaRepository<Author, Integer> {
     List<Author> findByPaper(ResearchPaper paper);
 
     // Find all author entries for a specific user
-    List<Author> findByUserId(Integer userId);
+    List<Author> findByUserId(Long userId);
 }
